perubahan dan penambahan layout, penambahan untuk controller dan database

+++ database/migrations/2025_11_06_000000_create_waste_data_table.php
@@ -6,22 +6,16 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::create('waste_data', function (Blueprint $table) {
             $table->id();
             $table->string('category');
-            $table->unsignedInteger('price_per_kg');
+            $table->decimal('price_per_kg', 15, 2);
             $table->timestamps();
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('waste_data');
